move Session type to legacy module
