+++ Cargo.toml
@@ -55,6 +55,7 @@ crossterm = { version = "0.23.2", features = ["event-stream"] }
 qrcode = { version = "0.12.0", optional = true }
 gethostname = "0.4.3"
 secrecy = { version = "0.8", features = ["serde"] }
+zeroize = "^1.4.3"
 
 [dev-dependencies]
 tempdir = "0.3"
